+++ backend/pkg/buildinfo/buildinfo_test.go
@@ -1,0 +1,9 @@
+package buildinfo
+
+import "testing"
+
+func TestCurrentUsesBuildVariables(t *testing.T) {
+	if Current() != (Info{Version: Version, Revision: Revision}) {
+		t.Fatalf("Current() = %#v", Current())
+	}
+}